-diarizer (CNAME); fedora→diarizer, +fedora
I didn't need a CNAME for diarizer.nono.io; I own diarizer.com which is good
enough.

I want a Fedora box that I can use to stage my deployments (I can't use tara
anymore because I've decided not to leave it on all the time).

So I decided to rename the diarizer.com VM from "fedora.nono.io" to
"diarizer.com". And do a migration to a different datastore so that the files
are renamed, too.

And once the move is complete I'm going to create a new VM "fedora.nono.io",
which will be my base of operations for redploying BOSH.