Grand nameserver (NS) reshuffling
We split `sslip.io`'s nameservers from the other domains. Previously, the
sslip.io nameservers did double-duty: they used a pipe (BASH) backend to
provide DNS for sslip.io, and they used a BIND backend for the other domains.

I plan to swap in a custom nameserver (not PowerDNS) for the `sslip.io` domain,
and it doesn't have the ability to act as a secondary (slave) server for the
other domains. So I am forced to re-organize my nameservers prior to
implementing the custom nameserver.

- All the BIND nameservers manage the following domains:
  - diarizer.com
  - foundry.fun
  - nono.com
  - nono.io
- And use the following nameservers:
  - ns-digitalocean.nono.io
  - ns-he.nono.io

- The PowerDNS nameservers manage the following domain:
  - sslip.io
- And use the following nameservers:
  - ns-aws.nono.io
  - ns-azure.nono.io
  - ns-gce.nono.io

To test DNS/whois: <http://www.webdnstools.com/dnstools/domain_check>