meet.cunnie.com: Let's Encrypt, nginx
https://meet.cunnie.com is a vanity URL for our ongoing family Google Meet.

A 301 Redirect is the answer, a simple nginx configuration.

Also the URL _must_ be HTTPS for it to work on Android phones; we use Let's
Encrypt.

We removed the nginx configuration for the no-longer-needed buzzer.nono.io.

The changes aren't as large as they might appear; we took the opportunity to
reformat the nginx configuration.